Fixed Race Condition when registring CrudServices
The way the services were registered in Services allowed the process to
be interuppted when two threads were requesting a service for the 1st
time at the same moment. Now, we register all services at application
startup and pace them in a final map. When retrieving a service just
read it from the map & build a service only if its really missing.

This issue is already fixed in upstream version 2.0.0, but that requires
a newer version of Spring than used in my project.